Third submission
1. Moved from deprecated code (onActivityResult() and onRequestPermissionResult()) to the latest recommended alterative (ActivityResultContracts)

2. The device location settings and the required permissions are properly handled

Google recommends to display one permission request at a time so I chained the requests. If the location service is disabled or a permission is denied, a warning message is displayed and the app do not ask for other required permissions.

For example, if the location service is turned off, the app does not request the location permissions.

In SaveReminderFragment.kt the  requests chain is:
checkIfLocationEnabled -> requestAccessFineLocationPermission -> requestBackgroundPermission

If any permission is missing, a reminder is saved but a geofence is not created.

In SelectLocationFragment.kt the requests chain is:
checkIfLocationEnabled -> checkFineLocationPermission

If any permission is missing, the user location is not automatically detected and the “my location” layer is disabled.

4. Added “check_loading” test.

5. Captured a video walkthrough and added it in the readme file.

6. Minor fixes (clear a few app inspection warnings)