Sixth submission
1. Directly display the device location settings dialog to the user instead of openning the settings dialog.

2. Use location updates.
In some cases fusedLocationProviderClient.lastLocation may be null.
One such case is right after turning on the location service. The location is not yet resolved and the location cache is empty  so the lastLocation is null.
If lastLocation is null, the app starts requesting location updates. When the first location is received, the app centers the map on this location and it stops requesting location updates.

3. Added flowcharts for SaveReminderFragment and SelectLocationFragment in readme.md.
The most important functionality of these classes is displayed in these flowcharts.